// frontend/src/adapter/api/__generated/api.ts
export interface Item {
    /**
     * 
     * @type {string}
     * @memberof Item
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof Item
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof Item
     */
    'description'?: string;
}
// ...
export interface ShoppingList {
    /**
     * 
     * @type {string}
     * @memberof ShoppingList
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof ShoppingList
     */
    'name': string;
    /**
     * 
     * @type {string}
     * @memberof ShoppingList
     */
    'description'?: string;
    /**
     * 
     * @type {string}
     * @memberof ShoppingList
     */
    'createdAt': string;
    /**
     * 
     * @type {boolean}
     * @memberof ShoppingList
     */
    'isFavorite': boolean;
    /**
     * 
     * @type {Array<ShoppingListShoppingListItemsInner>}
     * @memberof ShoppingList
     */
    'shoppingListItems'?: Array<ShoppingListShoppingListItemsInner>;
}
// ...
export interface ShoppingListShoppingListItemsInner {
    /**
     * 
     * @type {boolean}
     * @memberof ShoppingListShoppingListItemsInner
     */
    'isPurchased'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof ShoppingListShoppingListItemsInner
     */
    'quantity'?: number;
    /**
     * 
     * @type {Item}
     * @memberof ShoppingListShoppingListItemsInner
     */
    'item'?: Item;
}